+++ ArgPrograma40Clase7/src/main/java/com/mycompany/argprograma40clase7/ItemCarrito.java
@@ -19,7 +19,7 @@ public class ItemCarrito {
         this.cantidad = cantidad;
         this.producto = producto;
     }
-    public int precio(){
+    public Double precio(){
         return cantidad * producto.getPrecio();
     }
 }
